+++ CHANGELOG.md
@@ -12,12 +12,18 @@ and this project adheres to [Semantic Versioning](https://github.com/AldaronLau/
  - `From<Audio> for Box<[i16]>` impl
  - `Audio::as_slice` method
  - `Audio::as_slice_mut` method
+ - `Audio::with_i8_buffer` function
+ - `Audio::with_i16_buffer` function
+ - `Audio::with_f32_buffer` function
+ - `Audio::with_f64_buffer` function
 
 ### Removed
  - `From<Audio> for Box<[u8]>` impl
  - `From<Audio> for Box<[u16]>` impl
  - `Audio::as_u8_slice` method
  - `Audio::as_u8_slice_mut` method
+ - `Audio::with_u8_buffer` function
+ - `Audio::with_u16_buffer` function
 
 ## [0.1.0] - 2020-08-15
 ### Added